modified integer datatypes

+++ src/main/resources/nl/clockwork/mule/ebms/hsqldb/hsqldb.sql
@@ -9,9 +9,9 @@ CREATE TABLE ebms_message
 	time_stamp				TIMESTAMP				NOT NULL,
 	cpa_id						VARCHAR(256)		NOT NULL,
 	conversation_id		VARCHAR(256)		NOT NULL,
-	sequence_nr				INTEGER					NULL,
+	sequence_nr				SMALLINT				NULL,
 	message_id				VARCHAR(256)		NOT NULL,
-	message_nr				INTEGER					DEFAULT 0 NOT NULL,
+	message_nr				SMALLINT				DEFAULT 0 NOT NULL,
 	ref_to_message_id	VARCHAR(256)		NULL,
 	time_to_live			TIMESTAMP				NULL,
 	from_role					VARCHAR(256)		NULL,
@@ -20,7 +20,7 @@ CREATE TABLE ebms_message
 	service						VARCHAR(256)		NOT NULL,
 	action						VARCHAR(256)		NOT NULL,
 	content						CLOB						NULL,
-	status						INTEGER					NULL,
+	status						SMALLINT				NULL,
 	status_time				TIMESTAMP				NULL,
 	PRIMARY KEY (message_id,message_nr)
 );
@@ -28,7 +28,7 @@ CREATE TABLE ebms_message
 CREATE TABLE ebms_attachment
 (
 	message_id				VARCHAR(256)		NOT NULL,
-	message_nr				INTEGER					NOT NULL,
+	message_nr				SMALLINT				NOT NULL,
 	name							VARCHAR(256)		NULL,
 	content_id 				VARCHAR(256) 		NOT NULL,
 	content_type			VARCHAR(255)		NOT NULL,
@@ -40,8 +40,8 @@ CREATE TABLE ebms_event
 (
 	message_id				VARCHAR(256)		NOT NULL,
 	time							TIMESTAMP				DEFAULT NOW() NOT NULL,
-	type							INTEGER					NOT NULL,
-	status						INTEGER					DEFAULT 0 NOT NULL,
+	type							SMALLINT				NOT NULL,
+	status						SMALLINT				DEFAULT 0 NOT NULL,
 	status_time				TIMESTAMP				DEFAULT NOW() NOT NULL,
 	uri								VARCHAR(256)		NULL,
 	error_message			CLOB						NULL,
